Begin using the virtual frame for assignment statements of the form:
  <local> = <literal>

Literals are pushed on the frame as virtual elements.  Assignment of a
virtual element to a frame slot (eg, local or parameter) makes that slot
virtual too.  The above code compiles into a single move of a constant to
memory, and that move is actually only caused by spilling the frame.

Visiting expressions is guarded by spilling the frame to keep everything
spilled in code that assumes it.  That's ugly but temporary.
Review URL: http://codereview.chromium.org/10993