Experimental: thread the virtual frame through the deferred code generated at the end of functions.  It is piggybacked on the jump target infrastructure pairing the label of a block with the epected frame on entry.
This is simple and works for now but eventually we want to get the code
to make a frame mergable off the fast path or do away with it
completely.

The inlined code for the smi part of binary operations is duplicated,
but that is a temporary situation.
Review URL: http://codereview.chromium.org/13339