Experimental: begin using the register allocator for do and while loops, as well as continue and break statements.
This required fixing branches to blocks that have an expected frame.
The expected frame needs to be split (cloned), with one copy being
merged to the expected frame and the other flowing to the fall
through.

It also requires reestablishing a spilled frame when a break or
continue jumps into spilled code.
Review URL: http://codereview.chromium.org/17044