This is really Kasper's change that I took because Kasper is unavailable to work on it today.
Combine smi and overflow checks for count operations.  This generates
more compact code because we are always using big encodings of jumps
and also it has the effect of spacing jumps.

This can be done for other overflow and smi checks.  I'll do that
separately.


Review URL: http://codereview.chromium.org/17611