Fix disassembling of set instructions.
I don't really like the pointer to member thing, but it was the best I
could think of to not replicate the code.

We could use the NameOfByteRegister and PrintRightByteOperand in a
number of other places (such as test_b, but I will leave that for
another changelist).
Review URL: http://codereview.chromium.org/18313