Experimental: add a function to the API for recording assembler positions to record the start of each compiled function.
This seemed easiest, since the deferred code expects a statement
position and there is not necessarily a statement in the body to pass
to CodeForStatement.

Since it's on the branch, it is not ported to the ARM but will be
before we merge back to bleeding_edge.
Review URL: http://codereview.chromium.org/18592