Experimental: improve the code generated for possibly boolean-valued expressions.
Instead of branching to the true control target and jumping to the
false one, track which one is the preferred fall-through.  Use this to
branch to the non-fall-through and bind the fall-through.  This avoids
a jump to the next instruction or else a branch around a jump.

It becomes more difficult to tell if we have compiled an expression
into control flow or not.
Review URL: http://codereview.chromium.org/21077