Experimental: allow the virtual frame to explicitly indicate sharing of memory (as well as register) elements.
A new type of frame element is introduced, which is a copy of a memory
or register element.  As a simple first implementation, sets of copies
are backed by their lowest frame element.  This requires adjusting the
backing store when moving elements up and down in the frame.

Copied elements have copy-on-write semantics---modifying the backing
frame element requires finding the appropriate new backing element and
updating all copies in the set.

Review URL: http://codereview.chromium.org/19755