Experimental: remove VirtualFrame::MakeMergable, which munged up a frame by removing constants and copies from it.
Instead, compute an entry frame for a basic block when the block is
bound.  The entry frame is computed based on the frames that reach it
via forward edges (ie, that have reached it at binding time).  The
entry frame is allowed to contain constants and copies unless the jump
target is bidirectional.

Review URL: http://codereview.chromium.org/21292