Add an option to the SCons build to control Microsoft Visual C++ link-time code generation.
Turning off Microsoft Visual C++ link-time code generation makes linking on Windows significantly which is convenient during development. There has also been a request from embedders for a way to control this for embedding application which does not use link-time code generation.

The name of the option is msvcltcg and the default value is 'on' to keep the same behaviour as before.
Review URL: http://codereview.chromium.org/69026