Change to limit the amount of space we waste due to fragmentation in old space. Before this change we would compute the fragmentation limit before the GC, but that means that we take all the garbage into account - and consequently we almost never hit the limit (because we have a lot of garbage).
This change changes the policy to compact on the *next* GC if we
determine that we have too much fragmentation after doing a GC.
We'll do a GC if we've wasted more than 1MB and more than 15% of
the old space size.

This can be further improved by computing whether or not to compact
during the marking phase.
Review URL: http://codereview.chromium.org/202008