Added ES5 conform tests for getOwnPropertyNames.
Most of the test cases fail as the different objects according to the tests 
can not have additional properties attached to them. I will file a bug report
on the es5 conform site as they should allow this. Some of the test fails
because we still miss some of the es5 features used.


Review URL: http://codereview.chromium.org/545109