Fix dtoa for the LP64 model on 64-bit systems, where long is 64 bits.
dtoa.c:69 states: #define Long int on machines with 32-bit ints and 64-bit
longs. This change was made in Chromium's version of dtoa in r21901. I don't
know why this apparently hasn't caused a problem in v8 until now. I was
unable to build a snapshot on x86_64 Mac without this change.

BUG=706
TEST=x86_64 Mac GYP/Xcode-based Chromium build (still depends on other changes)