- Make ROOT_ACCESSOR, STRUCT_MAP_ACCESSOR, SYMBOL_ACCESSOR macro-generated   methods on Heap non-static. - Inline the Heap class into Isolate. - Fix Isolate::Init to set global_isolate to the isolate being initialized. - Define the macro HEAP to use during conversion.
HEAP expands into v8::internal::Isolate::current()->heap(). Since we haven't 
finalized how the heap (or parts of the heap) will be accessed from inside V8, 
using the macro allows us to make progress without worrying about reformatting 
code ("Heap::" is the same length as "HEAP->"). In addition, for those methods 
where we decide to keep a local copy of the pointer to the current heap, HEAP-> 
can become heap-> again without changing formatting.

THIS is defined within heap.cc to be equivalent to HEAP. THIS will go away as 
methods that are morally non-static but are currently marked as static are 
converted.

Review URL: http://codereview.chromium.org/2435001